Allow http protocol in storageAPIHost
