+++ internal/pkg/service/cli/dependencies/remote.go
@@ -5,13 +5,10 @@ import (
 
 	"github.com/keboola/go-client/pkg/keboola"
 
-	"github.com/keboola/keboola-as-code/internal/pkg/filesystem"
-	projectManifest "github.com/keboola/keboola-as-code/internal/pkg/project/manifest"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/cli/event"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/cli/options"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/common/dependencies"
 	"github.com/keboola/keboola-as-code/internal/pkg/utils/errors"
-	"github.com/keboola/keboola-as-code/internal/pkg/utils/strhelper"
 )
 
 // remote dependencies container implements Project interface.
@@ -55,34 +52,6 @@ func newProjectDeps(ctx context.Context, cmdPublicDeps ForLocalCommand, opts ...
 		Project:         projectDeps,
 		eventSender:     eventSender,
 	}, nil
-}
-
-func storageAPIHost(fs filesystem.Fs, opts *options.Options) (string, error) {
-	var host string
-	if fs.IsFile(projectManifest.Path()) {
-		// Get host from manifest
-		m, err := projectManifest.Load(fs, true)
-		if err != nil {
-			return "", err
-		} else {
-			host = m.APIHost()
-		}
-	} else {
-		// Get host from options (ENV/flag)
-		host = opts.GetString(options.StorageAPIHostOpt)
-	}
-
-	// Fallback
-	if host == "" {
-		host = "connection.keboola.com"
-	}
-
-	// Validate host
-	if host = strhelper.NormalizeHost(host); host == "" {
-		return "", ErrMissingStorageAPIHost
-	} else {
-		return host, nil
-	}
 }
 
 func (r *remote) EventSender() event.Sender {
