+++ pkg/lib/operation/project/local/template/use/operation.go
@@ -274,7 +274,7 @@ func (p *TemplatePlan) Invoke(ctx context.Context) (*Result, error) {
 		inputValuesMap := p.options.Inputs.ToMap()
 		for inputName, cKey := range tmplCtx.InputsUsage().OAuthConfigsMap() {
 			if len(inputValuesMap[inputName].Value.(map[string]interface{})) == 0 {
-				oauthWarnings = append(oauthWarnings, fmt.Sprintf("- https://%s/admin/projects/%d/components/%s/%s", p.deps.StorageAPIHost(), p.deps.ProjectID(), cKey.ComponentID, cKey.ID))
+				oauthWarnings = append(oauthWarnings, fmt.Sprintf("- %s/admin/projects/%d/components/%s/%s", p.deps.StorageAPIHost(), p.deps.ProjectID(), cKey.ComponentID, cKey.ID))
 			}
 		}
 		if len(oauthWarnings) > 0 {
